+++ src/edu/umass/cs390cg/atmosphere/geom/shapes/Shape.java
@@ -4,5 +4,5 @@ import edu.umass.cs390cg.atmosphere.geom.HitRecord;
 import edu.umass.cs390cg.atmosphere.geom.Ray;
 
 public interface Shape {
-  HitRecord hit(Ray ray, float tmin, float tmax);
+  HitRecord hit(Ray ray, double tmin, double tmax);
 }
